Added delegate function with amount spec
Added a delegate function where we can specify the amount we want to delegate , <code>delegate_amount & delegateBySig_amount</code>
This is usefull in cases where the tokens are locked in a timelock contract but we still want to allow the holders to vote (even if the tokens are only promised to the address, for example after a 3 year lock, and are yet to actually be owned by their future owner). In this case, the contract can have a certain amount of tokens in its balance and still only give a portion of the votes to each user.
Maybe it is easier to join the functions into only one, since the total balance of the caller of the address is always simpler to get via code, and so it can be placed in the "amount variable". This would save some gas